Move routes out of index to routes folder. Validate email input at registration.
