+++ package.json
@@ -22,6 +22,7 @@
     "passport": "^0.6.0",
     "passport-local": "^1.0.0",
     "superagent": "^8.1.2",
+    "validator": "^13.11.0",
     "winston": "^3.10.0"
   }
 }
